+++ lib/drivers/uart_stdio.h
@@ -22,4 +22,4 @@ uart_t uart_stdio_init(uint32_t baud);
 void uart0_write_byte(int8_t b);
 int8_t uart0_read_byte_blocking(void);
 
-uint8_t gets_nonblocking(char *buffer, uint8_t max_length);
+uint16_t gets_nonblocking(char *buffer, uint16_t max_length);
